requirements.txt
# MkDocs and core dependencies
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Blog functionality
mkdocs-material[blog]

# Additional plugins
mkdocs-git-revision-date-localized-plugin>=1.2.0
mkdocs-publisher>=1.4.0

# Development and testing dependencies
pytest>=7.0.0
yamllint>=1.32.0
pre-commit>=3.0.0

# Optional but recommended
linkchecker>=10.0.0
safety>=2.0.0